Module init.js is added
