+++ main.js
@@ -15,7 +15,7 @@
 
 import { addComment, getAllComments } from "./modules/api.js";
 import { getTime, disableButton, enableButton } from "./modules/utilitities.js";
-import { renderComments } from "./modules/renderComments.js";
+import { renderComments } from "./modules/renderElements.js";
 
 const removeCommentButtonElement = document.querySelector('#remove-comment-button');
 const addFormElement = document.querySelector('.add-form');
@@ -35,7 +35,8 @@ const fetchAndRenderComments = () => {
                 return {
                     name: comment.author.name,
                     date: getTime(new Date(comment.date)),
-                    text: comment.text,
+                    text: comment.text.replaceAll("<", "&lt;").replaceAll(">", "&gt;")
+                        .replaceAll('QUOTE_BEGIN', '<div class="quote">').replaceAll('QUOTE_END', '</div>'),
                     isLiked: comment.isLiked,
                     likes: comment.likes
                 }
@@ -56,83 +57,6 @@ const fetchAndRenderComments = () => {
 
 // Вызов функции с GET-запросом для получения списка комментариев из API
 fetchAndRenderComments();
-
-// Инициализация обработчика события для кнопок лайков: при нажатии на пустое сердечко оно закрашивается и 
-// счетчик увеличивается на единицу, при нажатии на закрашенное сердечко оно становится пустым и счетчик уменшается на единицу.
-const initLikeButtonListener = () => {
-    const likeButtonElements = document.querySelectorAll('.like-button');
-
-    for (const likeButtonElement of likeButtonElements) {
-        likeButtonElement.addEventListener('click', (e) => {
-            e.stopPropagation();
-            const index = likeButtonElement.dataset.index;
-
-            likeButtonElement.classList.add('-loading-like');
-
-            delay(2000)
-                .then(() => {
-                    comments[index].likes = comments[index].isLiked
-                        ? comments[index].likes - 1
-                        : comments[index].likes + 1;
-                    comments[index].isLiked = !comments[index].isLiked;
-                    likeButtonElement.classList.remove('-loading-like');
-                    renderComments({ comments });
-                });
-        });
-
-    }
-}
-
-// Инициализация обработчика событий для кнопок редактирования комментария: при нажатии на кнопку "Редактировать" текст комментария заменяется
-// на поле ввода, в которое подставлен текущий текст комментария, а кнопка "Редактировать" заменяется на кнопку "Сохранить".
-const initEditButtonListener = () => {
-    const editButtonElements = document.querySelectorAll('.edit-comment-button');
-
-    for (const editButtonElement of editButtonElements) {
-        editButtonElement.addEventListener('click', (e) => {
-            e.stopPropagation();
-            const index = editButtonElement.dataset.index;
-            if (comments[index].isEdit === false) {
-                comments[index].isEdit = true;
-            }
-            renderComments();
-        });
-    }
-}
-
-// Инициализация обработчика событий для кнопок сохранения отредактированного комментария: при нажатии на кнопку "Сохранить" поля ввода заменяется
-// на отредактированный комментарий, который сохраняется в массиве JS, а кнопка "Сохранить" заменяется на кнопку "Редактировать".
-const initSaveButtonListener = () => {
-    const saveButtonElements = document.querySelectorAll('.save-comment-button');
-
-    for (const saveButtonElement of saveButtonElements) {
-        saveButtonElement.addEventListener('click', (e) => {
-            e.stopPropagation();
-            const index = saveButtonElement.dataset.index;
-            if (comments[index].isEdit === true) {
-                comments[index].isEdit = false;
-                comments[index].text = document.querySelectorAll('.edit-comment-form-text')[index].value;
-            }
-            renderComments();
-        });
-    }
-}
-
-// Инициализация обработчика события по клику на комментарий для ответа на комментарий: 
-// не должен срабатывать при нажатии на лайк и на кнопки "Редактировать/Сохранить"
-const initCommentReplyListener = () => {
-    const commentElements = document.querySelectorAll('.comment');
-
-    for (const commentElement of commentElements) {
-        commentElement.addEventListener('click', (e) => {
-            if (e.target.classList.contains('edit-comment-form-text')) {
-                return;
-            }
-            const index = commentElement.dataset.index;
-            commentInputElement.value = `QUOTE_BEGIN ${comments[index].name}: \n ${comments[index].text}QUOTE_END \n \n`;
-        });
-    }
-}
 
 // Рендер-функция, которая отрисовывает форму ввода комментрия.
 const renderInputBox = (name, comment) => {
@@ -273,12 +197,3 @@ removeCommentButtonElement.addEventListener('click', () => {
     comments.splice(index, 1);
     renderComments({ comments });
 });
-
-// Функция для имитации запросов в API
-function delay(interval) {
-    return new Promise((resolve) => {
-        setTimeout(() => {
-            resolve();
-        }, interval);
-    });
-}
